constructing the call graph
for each qgl function, note which functions it calls

Right now we assume that each function name is unique
across a flat, contextless namespace, which only works
for a functional subset of python with function defs
in an effectively global scope.  We have to make assumptions
about the context of each call in order to do our analysis
instead of the usual "we'll figure this out from context at
runtime" way that Python resolves symbols.