add methods for collapsing names, getting fparams
collapsing names: turn the ast for something like 'a.b.c' back
into a string 'a.b.c', which is usually a nicer representation
(I am still on the fence in some contexts)

the _qbit_decl was moved from check_qbit to the importer,
so that it can be used wherever the importer is available.
it does not use any cached state, so it doesn't need to be
part of either class, except for the convenience of the
error messages.  I am not satisfied with this organization;
we need a general set of helpers we can mix in with the
different classes instead of lugging an importer around.