beginning of 'import-from'
half-baked; we need another level of indirection
because we can't just link the referenced methods
with the current context, because when we're
processing the referenced method they need to be
processed in the context they were originally defined
in...

So if you have

	from x import y

then when you expand y, you don't expand it in the
local namespace, you expand it in the context defined
by x.

The current code expands y in the current context,
which isn't what we want.