move away from the custom visitors
Now that we're annotating every node with the name
of the source file it came from, there's no need for
the visitor mix-ins for pretty-printing error messages.

For the short term I'm leaving the mix-ins in place
(but reimplementing them in terms of module methods)
but eventually they should be removed.

TODO: move the state out of a module variable so that
it doesn't have to appear "global".  It can be a
class variable instead.