adding module from meta, with bugfixes
NOTE: this code is BSD licensed.  Contains BBN project edits
but we are not the original copyright holders

the 'meta' package code for pretty-printing ASTs
doesn't understand Python3 'with' statements;
this version fixes that problem.

There are allegedly other shortcomings in this
module, which we'll either need to fix, or find
new replacement for.