squash multiple repetitions of the same warning
if we inline a function more than once (or later,
unroll loops, etc) then we could end up detecting
things that look bogus more than once, and printing
identical warnings/errors (that all reference the
same line in the original source file), and this
is not useful.  So track all the messages generated,
and only emit each once.