recurse the inlining through called functions
even if we can't inline a function, we might be
able to inline calls made by that function, so
make an attempt to do so.

An open issue is how to represent and access the
transformed functions.  Right now they are added
as an annotation to the original definition of the
function, but that's a bit clunky.  We could also
add them to the namespace, but what to do with the
final code is awkward.  The problem is that we need
to get the new code somewhere that the interpreter
will find it when it's time to actually execute
the program.