starting new type analyzer
the new type analyzer is better at finding variables
and infering their types, but not perfect: it doesn't
know how to deal with complicates rvals or actual parameters,
and doesn't know what to do with tuple assignments yet

the intent of the new analyzer was to replace the ugly
old analyzer, which was getting too complicated to maintain.
with the new features, however, the new analyzer is getting
hairy as well.