add support for 3.5 AST
In 3.5 vs 3.4:

In a "Call":

- the "starargs" member of an ast.Call node is removed
    and this is expressed with an args entry of type "ast.Starred"

- the "kwargs" member of an ast.Call node is removed
    and this is expressed with a keywords entry of None, whose
    value is the value of the kwargs

In addiion to adding functionality to handle Starred objects
and deal with keywords of "None", also added logic to ignore
the starargs and kwargs members, if they're absent, so things
also work for 3.4.

There may be other differences, which we'll have to find
by trial-and-error (or considerably more complete testing).
The release notes for 3.5 do not mention AST changes, so
we'll have to discover them via inspection.