relax handling of import 'errors'
some imports are in the system libraries; we don't
process these libraries

some imports are implemented by binary files (.zip or .so,
and possibly other formats) and we don't process these because
we can parse them

So if we can't find the source code for a module, it doesn't
mean that it's a fatal error.  We make note of it and move on.