add 'repeat' to the grouper
if there's a repeat statement inside a concur block,
we break it down the repeat statement into one or more
repeat statements, one per QBIT partition.
This makes it easier to map directly to the hardware.

For example:

	with concur:
	    for x in range(2):
	        for q in [QBIT_1, QBIT_2]:
		    X90(q)

after the expansion/unrolling process is complete, this will
look something like:

	with concur:
	    with qrepeat(2):
		X90(QBIT_1)
		X90(QBIT_2

But when we try to "group" the "with qrepeat" block we'll
be symied because both qbits are referenced in the same block.
So we'll split the block:

	with concur:
	    with qrepeat(2):
		X90(QBIT_1)
	    with qrepeat(2):
		X90(QBIT_2)

and then we can can partition successfully:

	with concur:
	    with seq:
		with qrepeat(2):
		    X90(QBIT_1)
	    with seq:
		with qrepeat(2):
		    X90(QBIT_2)