do substitution on orig calls of inlined funcs
when we do variable substitution for qbits in a statement,
do the same substitution on any inlined-functions.  This
is done so we can keep track of qbits that were referenced
in the original code, pre-optimization, that might not be
referenced now because of inconsistencies in the QGL
API.  For example, init(q) gets inlined as Wait(), but
we need to know that that Wait() depends on the q from
the init(q) call -- this q is implicit in the Wait(),
but we need to know it.