fail if Wait()/Sync() undefined in qglmain scope
The compilation is very likely to fail if Wait and/or Sync
are not defined in a scope visible from the function
declared as qglmain, because these functions may be called
by introduced code.  Therefore if we detect that these
are not defined, we complain and fail right away, forcing
the programmer to add the proper imports.

This is a gross hack and we need a better solution, but
for now, at least we get a meaningful error message