pass the shared importer into SingleSequence
SingleSequence was trying to reconstruct the namespaces,
which actually caused them to be re-parsed, undoing the
work we'd done when we seeded the namespaces with definitions
for things like Wait and Sync.

There should be a way to locate the importer for a given
namespace without reconstructing it, but right now there
isn't.