move import_native into NameSpace class
begin fixing the issues related to importing natively
in parallel with importing as QGL2.  Still a lot of
rough edges, including handling relative imports in
a graceful way.  (We might decide that it's OK to just
punt on this for now, since where we use them right now
are places where it doesn't matter--it only matters if
we use them in actual "QGL2" code)