don't recursively substitute calls; just copy
we keep a reference back to the original call whenever
we do a substitution/specialization, in order to keep
a reference to the actual parameters in the call.
This is important if there are actual parameters in
the original call that are qbits that are not referenced
in the substitution.  The common example is "init(q)",
which is obviously related to channel q, but which
expands to "Wait(); Sync()", which no longer mentions q
explicitly.  Keeping a reference back to the original
call to init tells us the actual value of the formal
parameter q at the moment the call occurred.

The bug was that we were trying to do substitution
recursively on the original call as well.  This doesn't
seem to make sense, causes bugs during substitution,
and it's not obvious why I added this code.  So to
remove the bugs, I'm removing the recursion.