simplify native_eval; add native_exec
simplify native_eval by always converting AST to text
and then compiling that.  AST could appear in several
forms, and compile could get confused.

Fixed a bug in handling local bindings; the empty
dictionary was being confused with "no dictionary"
and additions were therefore being thrown away.
Fix is to compare the dictionary to None, not non-True
(because empty dictionaries test as "False")

native_exec is an extension that lets the compiler know
that it's getting a statement instead of an expression,
because statements and expressions have to be compiled
using different modes.  This still has some rough
edges on it (statements shouldn't return a value, but
we still do return a dummy value...  this should be fixed.