add native import for the "base" file
this is a bit of a hack, because the base file isn't
imported as the result of an import statement, so if
there's any error we don't know how to frame this
error in terms of an executable statement.  Fortunately
if there were going to be an error, it probably would
have already happened... but that's not certain.