improve temp; don't mangle qgl_fname on new nodes
improve how temp names are made: if we're making a temp
name that is based on another temp name, try to figure out
what the original name was, and use that instead.  For example,
a name like X might have been given a temp name of X___qgl2_tmp_001,
and then that would be given name of X___qgl2_tmp_001___qgl2_tmp_002.
In the new mode, the second name would be X___qgl2_tmp_002.  This
throws away the info that this was based on X___qgl2_tmp_001, but
makes it easier to see that it's based on "X", which is the info
that a reader would almost always find most important.

for the qgl_fname, we used to lexically mark the fname of certain
synthetic nodes with a name that indicated that the node was
synthetic, but this broke things later because we need the fname
to find the namespace in which the nodes should be evaluated.
We might still want to mark these nodes (someday) but doing so
by rewriting the fname has become a bad idea.