permit builtin functions
if we can't resolve a symbol, see if it's
a builtin before giving up.

The test for whether something is a builtin is,
for reasons that are not obvious at all, different
for the Python shell and the file interpreter.
This is worth noting because it leads to strange
bugs otherwise