add QubitPlaceholder; fix rewriter, redirector
QubitPlaceholder is a value that can be used in place
of an actual Qubit, just to carry information forward
about the underlying name of the channel.

the rewriter is extended to rewrite the orig_call info
(so that the qubit origins can be meaningfully tracked
if all we have is the name of the variable and not the
underlying name)

the redirector is updated so that if a value is a
QubitPlaceholder, the value is replaced in the AST
with an ast.Name for the underlying qbit (and not a
redirection to a reference to the QubitPlaceholder,
which is not useful to the compiler).

Eventually it would be nice to use real Qubits/Channels
instead of placeholders