fix whitespace bug in indentation level shifts
pysourcegen had a harmless but increasingly annoying
bug that resulted in extra whitespace being inserted
on *both* sides of an indented block, because the logic
for determining how much whitespace should be added
prior to a newline was broken ("don't indent" was treated
the same as "indent to the current level").  This meant
that blocks would be indented something like the following
(with leading and trailing spaces replaced by +:

++++for x in foo:++++
++++++++with bar:++++++++
++++++++++++if some_condition:++++++++++++

at any nontrivial degree of indentation, the trailing
whitespace started to interfere with viewing the code
on a screen with any reasonable width.  As our code
becomes more complicated, this passes the level annoying
and becomes a problem.