prototype of AddSequential working; needs testing
this should add barriers everywhere necessary to
enforce sequentiality where the programmer requested it.
(also a few redundant barriers, to make things easier
to read)

There's a lot left to test, however.

The code to add barriers for the concur and infunc blocks
can be rolled up into this, so that we get rid of all of
the "implicit" barriers in the same pass and never have
to look at a version of the code that has a mix of explicit
and implicit barriers.