basic functionality complete; still missing stmnts
handles the basic statements but doesn't have the templates
for all the statements we want to handle (i.e. while loops,
with statements, etc).  Most of the functionality will
occur already because of the default handlers for the
types that compose these statements, but things like
handling the items clause of with statements won't be
exactly right without a little more work.