fix bug in detecting variables in AugAssign
we try to figure out which variables are actually variable
(and which are constants) by detecting potential changes
in their bindings.  We overlooked AugAssign statements,
so some variables were getting treated as constants even
though they could be reassigned.

This needs more review because there might be other types
of assignment that we're not detecting correctly