fix qbit references after barriers added
we have two mechanisms that depend on tracking all of the
qbits referenced by a subtree of the AST: which qbits we
need to protect via a barrier in order to implement any
synchronization required by that subtree, and whether
a particular statement contains any code that needs to
execute on a channel processor.

In order to synchronize things across multiple channels,
we might add barriers around a statement that reference
qbits not referenced by that statement -- in which case
the parent statement may need to be updated as well
so that the barriers will be placed on all the right channels
when the code is grouped by channel.

For example, in a sequential context:

    for q in [q1, q2]:
	with concur:
	    foo(q)

which becomes something like:

    with Qfor:
	with Qiter: # ONLY references q1
	    with concur:
		foo(q1)
	with Qiter: # ONLY references q2
	    with concur:
		foo(q2)

then when we add the barriers (shown here in a simplified form)
needed to make this sequential:

    with Qfor:
	with Qiter:
	    Barrier(1, [q1, q2])
	    foo(q1)
	    Barrier(2, [q1, q2])
	with Qiter:
	    Barrier(3, [q1, q2])
	    foo(q2)
	    Barrier(4, [q1, q2])

Now we need to update the references on each Qiter so that
they both reference [q1, q2] so that all four barriers will
appear on both q1 and q2.

The way we do this in this commit is by returning a new
list of referenced qbits to caller whenever we add barriers
to a body of statements, so that the parent of the body can
be updated.

This lacks elegance and I suspect there's a better way,
either by using two variables or by being smarter about
how the grouping is done.