add "implicit" Barrier import; fix qbit marking
We weren't marking the qbits of the original call when
we inlined a function that took qbits as parameters and
turned them into sequences of statements that didn't
(which is a bit weird on the face of it, but some
methods take implicit qbit parameters so this can occur).
This meant that the grouper would fail to group some
expressions properly, and they'd get lost.

Also add an "implicit" import for Barriers (and
eventually perhaps other QGL1 functions) so that
the user doesn't need to import functions, like
Barrier, that get added by the preprocessor.  This
functionality will be completed in sequences.py;
this is just the setup for the real fix.