start to fix metaprogramming problems
if we have a function reference assigned to a local variable,
then before we can invoke that function we need to "convert"
the call into a call to the function, rather than an application
of the local variable.  For stubs, this is straightforward but
icky.  For qgl2decl functions, we need to inline the function
(and unroll it, etc) as if it had been there all along.
We currently handle stubs, but not qgl2decls, and native functions
are problematic as well.