remove error msg if we call a "library" function
if a qgl2 function called an arbitrary function defined in
the system libraries (i.e. something in numpy) then this would
be flagged as an error because we didn't know how to check
the type constraints of those functions.  This is bogus and
has beens softened to a warning, which is still quasi-bogus
because we believe that these library functions will be used
frequently, and getting warned about each call is silly.

This is a symptom of the way we ignore "library" symbols during
the processing of imports (so we don't try to deal with definitions
that are incomprehensible to our preprocessor).  This means that
we don't even keep track of whether these symbols are defined.
It would be better to do that, and check that here, than it is
to complain that we don't know whether there's a binding at all.