remove bogus check for functions we don't have
because the importer does not process "system" libraries,
there may be symbols that we don't know about in our local
concept of namespaces that are actually available in the
namespace of the executable.  Therefore we shouldn't balk
if we see a symbol we haven't seen defined, but instead
continue.  If the symbol is truly undefined, this will be
detected later.