may revert - remove check_symtab from main
checking the symbol table doesn't work for
call-by-reference functions: we can't check them
until we know what the reference points to, which
isn't until runtime.  These almost always fail
(and in a perfect world would *always* fail) in
check_symtab because the bindings don't exist.

check_symtab may be completely redundant at this
point, so this change might be permanent (and
the code can also be removed).  If we want to keep
check_symtab, then we need to fix it so it doesn't
consider these errors.