exec qgl2 imported modules against native_globals
we tried to import the bare minimum into each namespace,
but were were missing indirectly-referenced symbols.

For example, if you have file a.py that contains

from b import beta

@qgl2decl
def alpha():
    beta()

and b.py that contains

def gamma():
    something

@qgl2decl
def beta():
    gamma()

then the reference to "beta" would be loaded, but when beta
called gamma, gamma would not be found in the namespace created
to resolve symbols while executing beta.  This is fixed by making
all of the symbols defined in b available to this namespace,
whether or not it's obvious that they will be referenced.