force file close rather than letting the gc do it
The Python garbage collector closes files when their
references fall out of scope, but how this is implemented
is implementation-specific and the unittest framework
takes a (justifiably) dim view of assuming it will happen
promptly and therefore flags lazily-closed files as
potential resource leaks.  This commit adds an explicit
close to the opens in the importer.