reset errors in compileFunction; print parameters
if we call compileFunction repeatedly, we need to
reset the error state before each call, or else any
prior errors will cause the current call to fail
immediately.

also added a line to print out what is being compiled,
because in the tests there's no indication of where
the output from one run ends and the next begins,
which makes it hard to read.