when we add checks, mark it as a change
in some cases we weren't counting the addition of
runtime checks as a "change" to the code, which
meant that we would discard the changes (because
if the inliner fails to find any changes to make,
it doesn't keep its new copy of the code).  But
we want to save these changes, so we need to consider
them as first class changes.

This bug was subtle to find because the only code
that *only* has checker changes is code that is not
be transformed in any other way, which is a trivial
sequence of QGL commands in the qgl2main function
with no conditional statements, no concurrency, etc.