fix how call-by-ref builds its with-infunc
the parameters to the with-infunc for functions
passed by reference was built using the new
parameters for the expanded function, rather than
original parameters for the called-by-reference
function, which mean that later checks to make sure
that every variable was properly defined would
raise a warning.

For example, if we had something like:

    # Let X be the name of an actual function
    def X(a: int, b: int, c: int):
	something

    # Make foo be a reference to X
    foo = X

    foo(1, 2, 3)

The call to foo would be expanded to add the type checks.  (for
illustrative purposes, I'm expanding everything fully, which is
not always necessary, and certainly is not necessary in this
case.  In the future we'll try to reduce the unnecessary expansions.)

    a_tmp = 1
    b_tmp = 2
    c_tmp = 3
    a_ass = a_tmp
    b_ass = b_tmp
    c_ass = c_tmp
    QGLcheck()
    foo(a_ass, b_ass, c_ass)

and then foo itself would be expanded:

    a_tmp = 1
    b_tmp = 2
    c_tmp = 3
    a_ass = a_tmp
    b_ass = b_tmp
    c_ass = c_tmp
    QGLcheck()
    with infunc(foo, a_ass, b_ass, c_ass):
	something

and then the checks would be moved to be inside the with-infunc:

    a_tmp = 1
    b_tmp = 2
    c_tmp = 3
    with infunc(foo, a_ass, b_ass, c_ass):
	a_ass = a_tmp
	b_ass = b_tmp
	c_ass = c_tmp
	QGLcheck()
	something

And this would cause a problem because when we do a scope check on the
with-infunc, we'll see that it depends on a_ass, b_ass, and c_ass,
but none of those have been defined yet: they're defined within the
with block.  So the for the checks to work without complaining, we need
to rewrite the infunc to have the "original" parameters:

    ...
    with infunc(foo, a_tmp, b_tmp, c_tmp):
	...