fix call to compileFunction; preserve local state
compileFunction was being called with args out of order;
probably due to an undetectedly botched merge

also add a call to preserve the local state so that
it's not overwritten.  Ignoring the attempt to streamline
the local state by throwing away unreferenced variables,
because when we do that we tend to throw away things
that we're going to need in the future.