move pulse shapes out of RabiMin
the reference to a function depends on how it is imported
(in some cases) so in order to get the "same" reference to
the tanh function in the QGL2 and the QGL1 test code, we
need to move tanh into a different file than the QGL2 code.
If the two references aren't identical, then the unittests
will fail because the shapeParams will look different.

The reason for this is not clear, but this seems to solve
the problem.