don't crash on the empty program
If there are no qbit sequences whatsoever, don't crash.
Give the user a warning message telling them that compilation
has aborted due to having nothing to compile, but don't
treat this as an error per se because perhaps the program
was merely executed for its classical effect -- which is
pointless, although permissible.