fix name resolution; short-circuit sym checks
in the name way that we fixed resolution for ast.Name
nodes, fix the resolution for simple names by chasing
names through multiple levels of indirection, if needed

for the short-circuiting -- if an actual parameter is
already a reference to a name, then use than name instead
of creating a scratch name to ensure that all APs are
names.  This is the common case, this helps a lot.
(we could do something similar with other constants,
but it's more complicated and doesn't help as much)