fix call to sys.exit()
The docs say that passing a string to sys.exit() will cause the
string to be printed to stderr, but that is not what I see when
I run the program -- the string isn't emitted anywhere.

Changed the sys.exit() to a call to print() followed by a call to
sys.exit().  This is more consistent with how we handle other
errors.