If we cannot find a function's AST, assume it is not a QGL2 function.
The problem was when expanding assignments in the EvalTransfomer, we now compute
the function type of any Call in is_measurement() prior to attempting to
evaluate the statement. However, this was choking on things like:
lengths = np.linspace(0,10e-6,11)

because
1) the AST value is an Attribute, so the function name is another layer down.
2) import.resolve_sym() does not return the AST for linspace. We previously
threw an error in this case.

I fixed 1), but for 2) I just return NONQGL2 instead of throwing an error. Is
this what we want?

cc @dellard