Fix last few tests.
Marking syndrome test_syn_1 as expected failure because it uses Edges are that
not in the default channel library. A better fix is to improve the channel setup
accordingly.

test_84_5 is technically a change in behavior. But, whereas before we didn't
return a result function, now the result function simply produces an empty list.
This seems like acceptable behavior.