improved parameter list handling, but incomplete
this passes the unit tests, but is incomplete in terms
of checking whether all the required parameters are present,
no bogus keywork parameters have been inserted, and
it doesn't insert code to fill in default values yet
in a general form.

Also doesn't report errors correctly, and fills the screen
with debugging messages...  so needs a lot of cleaning up.