change check of *args/**kwargs to func, not call
It's not clear what check_call_params was actually trying
to check, but it seemed broken, so I removed it.
There might be some python feature I'm forgetting about
and I'll have to add this back--but not exactly as such
because it wasn't checking for the cases we can actually
handle

I added a new check, at the same spot in the code,
for the formal parameters of a function (to make sure
that a **kwargs isn't among them, because we don't
support that).