added new check_call_actuals
this new function checks whether the call uses *args
or **kwargs, and disallows either.  We might want to
add these someday, but we're not using them right now,
I doubt the code was entirely correct, and it's a bit
hairy, so I'd rather explicitly disallow them than let
people use a feature that probably doesn't work.