using new utility functions; creating bugs
now that I'm using the utility functions wherever
I need to map calls to formal parameters, I realized
that half the stub functions use **kwargs, so these
are mandatory (unless I want to rewrite all the stubs,
and edit them whenever the QGL primitives change...
which is absolutely the wrong approach).

So next, the return of **kwargs.