almost complete **kwargs integration; buggy
three of the unittests are crashing; at least one of the unit
that used to fail now succceeds.  More work to be done here...

We might not be able to completely unify everything because the
way that *args and **kwargs are handled is still clunky.  When
we do have an *args parameter, we should pass it carefully.