fix return value for "checkless" stubs
There was a case that was never exercised: a stub
with no parameters, and therefore no checks, would
confuse the evaluator.  We might not ever have a stub
without any parameters, but we might want to turn
off checks (since they are time-consuming), which would
lead to the same problem.